+++ packages/react-components/react-aria/src/activedescendant/useOptionWalker.ts
@@ -94,6 +94,5 @@ export function useOptionWalker<TListboxElement extends HTMLElement>(options: Us
   return {
     optionWalker,
     listboxCallbackRef: setListbox,
-    listboxRef,
   };
 }