+++ packages/react-components/react-combobox/src/components/Combobox/useInputTriggerSlot.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
+import type { ActiveDescendantImperativeRef } from '@fluentui/react-aria';
 import { mergeCallbacks, useEventCallback } from '@fluentui/react-utilities';
-import { ActiveDescendantImperativeRef } from '@fluentui/react-aria';
 import type { ExtractSlotProps, Slot, SlotComponentType } from '@fluentui/react-utilities';
 import { ArrowLeft, ArrowRight } from '@fluentui/keyboard-keys';
 import { useTriggerSlot, UseTriggerSlotState } from '../../utils/useTriggerSlot';