+++ packages/react-components/react-combobox/src/utils/useTriggerSlot.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { ActiveDescendantImperativeRef } from '@fluentui/react-aria';
+import { useSetKeyboardNavigation } from '@fluentui/react-tabster';
+import type { ActiveDescendantImperativeRef } from '@fluentui/react-aria';
 import { mergeCallbacks, slot, useEventCallback, useMergedRefs } from '@fluentui/react-utilities';
 import type { ExtractSlotProps, Slot, SlotComponentType } from '@fluentui/react-utilities';
 import { getDropdownActionFromKey } from '../utils/dropdownKeyActions';
@@ -133,6 +134,7 @@ function useTriggerKeydown(
     }
   };
 
+  const setKeyboardNavigation = useSetKeyboardNavigation();
   return useEventCallback((e: React.KeyboardEvent<HTMLInputElement> & React.KeyboardEvent<HTMLButtonElement>) => {
     const action = getDropdownActionFromKey(e, { open, multiselect });
     const activeOption = getActiveOption();
@@ -150,6 +152,8 @@ function useTriggerKeydown(
         e.preventDefault();
         break;
     }
+
+    setKeyboardNavigation(true);
 
     switch (action) {
       case 'First':