+++ packages/react-components/react-datepicker-compat/src/components/DatePicker/useDatePicker.tsx
@@ -113,7 +113,7 @@ export const useDatePicker_unstable = (props: DatePickerProps, ref: React.Ref<HT
     formatDate = defaultFormatDate,
     highlightCurrentMonth = false,
     highlightSelectedMonth = false,
-    initialPickerDate = new Date(),
+    initialPickerDate: initialPickerDateProp,
     inlinePopup = false,
     isMonthPickerVisible = true,
     maxDate,
@@ -134,6 +134,9 @@ export const useDatePicker_unstable = (props: DatePickerProps, ref: React.Ref<HT
     value,
     ...restOfProps
   } = props;
+
+  const initialPickerDate = React.useMemo(() => initialPickerDateProp ?? new Date(), [initialPickerDateProp]);
+
   const calendar = React.useRef<ICalendar>(null);
   const [focus, rootRef, preventFocusOpeningPicker, preventNextFocusOpeningPicker] = useFocusLogic();
   const [selectedDate, formattedDate, setSelectedDate, setFormattedDate] = useSelectedDate({
@@ -144,6 +147,7 @@ export const useDatePicker_unstable = (props: DatePickerProps, ref: React.Ref<HT
   const [open, setOpenState] = usePopupVisibility(props);
   const fieldContext = useFieldContext();
   const required = fieldContext?.required ?? props.required;
+  const defaultId = useId('datePicker-input');
   const popupSurfaceId = useId('datePicker-popupSurface');
 
   const validateTextInput = React.useCallback(
@@ -368,9 +372,23 @@ export const useDatePicker_unstable = (props: DatePickerProps, ref: React.Ref<HT
   });
   input.ref = useMergedRefs(input.ref, ref, rootRef);
 
+  // Props to create a semantic but non-focusable button on the element with the click-to-open handler
+  // Used for voice control and touch screen reader accessibility
+  const inputLabelledBy = props['aria-labelledby'];
+  const inputId = props.id ?? defaultId;
+  const iconA11yProps = React.useMemo(
+    () => ({
+      role: 'button',
+      'aria-expanded': open,
+      'aria-labelledby': inputLabelledBy ?? inputId,
+    }),
+    [open, inputLabelledBy, inputId],
+  );
+
   const contentAfter = slot.always(props.contentAfter || {}, {
     defaultProps: {
       children: <CalendarMonthRegular />,
+      ...iconA11yProps,
     },
     elementType: 'span',
   });
@@ -384,6 +402,7 @@ export const useDatePicker_unstable = (props: DatePickerProps, ref: React.Ref<HT
       'aria-haspopup': 'dialog',
       readOnly: !allowTextInput,
       role: 'combobox',
+      id: inputId,
     },
     elementType: Input,
   });