+++ packages/react-components/react-rating-preview/src/components/Rating/useRating.tsx
@@ -11,6 +11,9 @@ import type { RatingProps, RatingState } from './Rating.types';
 import { RatingItem } from '../../RatingItem';
 import { StarFilled, StarRegular } from '@fluentui/react-icons';
 
+const defaultIconFilled = <StarFilled />;
+const defaultIconOutline = <StarRegular />;
+
 /**
  * Create the state required to render Rating.
  *
@@ -24,8 +27,8 @@ export const useRating_unstable = (props: RatingProps, ref: React.Ref<HTMLDivEle
   const generatedName = useId('rating-');
   const {
     color = 'neutral',
-    iconFilled = <StarFilled />,
-    iconOutline = <StarRegular />,
+    iconFilled = defaultIconFilled,
+    iconOutline = defaultIconOutline,
     max = 5,
     name = generatedName,
     onChange,
@@ -67,6 +70,7 @@ export const useRating_unstable = (props: RatingProps, ref: React.Ref<HTMLDivEle
         {
           ref,
           children: rootChildren,
+          role: 'radiogroup',
           ...props,
         },
         ['onChange'],