+++ packages/react-components/react-portal/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-portal
 
-This log was last generated on Tue, 30 Jan 2024 23:16:54 GMT and should not be manually modified.
+This log was last generated on Tue, 06 Feb 2024 17:55:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.4.13](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal_v9.4.13)
+
+Tue, 06 Feb 2024 17:55:21 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-portal_v9.4.12..@fluentui/react-portal_v9.4.13)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.19.0 ([PR #30392](https://github.com/microsoft/fluentui/pull/30392) by beachball)
 
 ## [9.4.12](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal_v9.4.12)